--- screens/CommunityScreen.js
@@ -1,15 +1,0 @@
-import React from "react";
-import styled from "styled-components/native";
-
-const Container = styled.View``;
-const Title = styled.Text``;
-
-const CommunityScreen = () => {
-  return (
-    <Container>
-      <Title>CommunityScreen</Title>
-    </Container>
-  );
-};
-
-export default CommunityScreen;
