--- screens/ProfileScreen.js
@@ -1,15 +1,0 @@
-import React from "react";
-import styled from "styled-components/native";
-
-const Container = styled.View``;
-const Title = styled.Text``;
-
-const ProfileScreen = () => {
-  return (
-    <Container>
-      <Title>ProfileScreen</Title>
-    </Container>
-  );
-};
-
-export default ProfileScreen;
